Workaround for conv_resol when pandas installed but not wanted (#79)
* Comment out print() call in create_working_order()

As title. It appears to have been accidentally left in

* fix PEP8 Errors (#68)

Fix errors raised by flake8

* Account Subscription Fix

The sample script fails to subscribe to account updates due to the items kwarg not being a list. 

It returns: ERROR 21 unrecognized group name

* Workaround for conv_resol when pandas installed but not wanted

This is a workaound for an edge case whereby a user might have Panda's installed but does not want to use it with the API wrapper. 

If you pass a resolution in accordance with the API guidelines e.g MINUTE, a ValueError will be raised.